Revert unneeded changes
